refactor context utils
