+++ apps/src/codebridge/codebridgeContext/index.ts
@@ -1,2 +1,1 @@
 export * from './codebridgeContext';
-export * from './utils';
